Handle setting of image block href when converting from classic to blocks
The strategy chosen is to always set 'Link To' to 'Custom URL' if an anchor tag
exists in the classic markup. Trying to detect 'media file' or
'attachment page' could potentially be too error prone, this approach at least
ensures the image's anchor link is the same in the resulting post.